Refactor and move the navigation code to the CoreFamilyProfileActivity for reuse in HF
