+++ opensrp-chw/src/main/java/org/smartregister/chw/activity/FamilyProfileActivity.java
@@ -2,20 +2,15 @@ package org.smartregister.chw.activity;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.support.v4.view.ViewPager;
-import android.view.View;
 
 import org.smartregister.chw.anc.activity.BaseAncMemberProfileActivity;
-import org.smartregister.chw.anc.domain.MemberObject;
 import org.smartregister.chw.core.activity.CoreAboveFiveChildProfileActivity;
 import org.smartregister.chw.core.activity.CoreChildProfileActivity;
 import org.smartregister.chw.core.activity.CoreFamilyProfileActivity;
 import org.smartregister.chw.core.activity.CoreFamilyProfileMenuActivity;
 import org.smartregister.chw.core.activity.CoreFamilyRemoveMemberActivity;
-import org.smartregister.chw.core.utils.ChildDBConstants;
-import org.smartregister.chw.core.utils.CoreChildUtils;
 import org.smartregister.chw.core.utils.CoreConstants;
 import org.smartregister.chw.fragment.FamilyProfileActivityFragment;
 import org.smartregister.chw.fragment.FamilyProfileDueFragment;
@@ -24,18 +19,13 @@ import org.smartregister.chw.model.FamilyProfileModel;
 import org.smartregister.chw.pnc.activity.BasePncMemberProfileActivity;
 import org.smartregister.chw.presenter.FamilyProfilePresenter;
 import org.smartregister.commonregistry.CommonPersonObject;
-import org.smartregister.commonregistry.CommonPersonObjectClient;
 import org.smartregister.family.adapter.ViewPagerAdapter;
 import org.smartregister.family.fragment.BaseFamilyProfileDueFragment;
 import org.smartregister.family.fragment.BaseFamilyProfileMemberFragment;
 import org.smartregister.family.util.Constants;
-import org.smartregister.family.util.DBConstants;
-import org.smartregister.family.util.Utils;
 import org.smartregister.view.fragment.BaseRegisterFragment;
 
 import java.util.HashMap;
-
-import static org.smartregister.chw.anc.util.Constants.ANC_MEMBER_OBJECTS.TITLE_VIEW_TEXT;
 
 public class FamilyProfileActivity extends CoreFamilyProfileActivity {
     private FamilyProfileActivityFragment profileActivityFragment;
@@ -120,83 +110,6 @@ public class FamilyProfileActivity extends CoreFamilyProfileActivity {
         }
 
         return viewPager;
-    }
-
-    public void goToProfileActivity(View view, Bundle fragmentArguments) {
-        if (view.getTag() instanceof CommonPersonObjectClient) {
-            CommonPersonObjectClient commonPersonObjectClient = (CommonPersonObjectClient) view.getTag();
-            String entityType = Utils.getValue(commonPersonObjectClient.getColumnmaps(), ChildDBConstants.KEY.ENTITY_TYPE, false);
-            if (CoreConstants.TABLE_NAME.FAMILY_MEMBER.equals(entityType)) {
-                if (isAncMember(commonPersonObjectClient.entityId())) {
-                    goToAncProfileActivity(commonPersonObjectClient, fragmentArguments);
-                } else if (isPncMember(commonPersonObjectClient.entityId())) {
-                    gotToPncProfileActivity(commonPersonObjectClient, fragmentArguments);
-                } else {
-                    goToOtherMemberProfileActivity(commonPersonObjectClient, fragmentArguments);
-                }
-            } else {
-                goToChildProfileActivity(commonPersonObjectClient, fragmentArguments);
-            }
-        }
-    }
-
-    public void goToOtherMemberProfileActivity(CommonPersonObjectClient patient, Bundle bundle) {
-        Intent intent = new Intent(this, getFamilyOtherMemberProfileActivityClass());
-        if (bundle != null) {
-            intent.putExtras(bundle);
-        }
-        intent.putExtra(Constants.INTENT_KEY.BASE_ENTITY_ID, patient.getCaseId());
-        intent.putExtra(CoreConstants.INTENT_KEY.CHILD_COMMON_PERSON, patient);
-        intent.putExtra(Constants.INTENT_KEY.FAMILY_HEAD, getFamilyHead());
-        intent.putExtra(Constants.INTENT_KEY.PRIMARY_CAREGIVER, getPrimaryCaregiver());
-        startActivity(intent);
-    }
-
-    public void goToChildProfileActivity(CommonPersonObjectClient patient, Bundle bundle) {
-        String dobString = Utils.getDuration(Utils.getValue(patient.getColumnmaps(), DBConstants.KEY.DOB, false));
-        Integer yearOfBirth = CoreChildUtils.dobStringToYear(dobString);
-        Intent intent;
-        if (yearOfBirth != null && yearOfBirth >= 5) {
-            intent = new Intent(this, getAboveFiveChildProfileActivityClass());
-        } else {
-            intent = new Intent(this, getChildProfileActivityClass());
-        }
-        if (bundle != null) {
-            intent.putExtras(bundle);
-        }
-        intent.putExtra(CoreConstants.INTENT_KEY.IS_COMES_FROM_FAMILY, true);
-        intent.putExtra(Constants.INTENT_KEY.BASE_ENTITY_ID, patient.getCaseId());
-        intent.putExtra(org.smartregister.chw.anc.util.Constants.ANC_MEMBER_OBJECTS.MEMBER_PROFILE_OBJECT, new MemberObject(patient));
-        startActivity(intent);
-    }
-
-    public void goToAncProfileActivity(CommonPersonObjectClient patient, Bundle bundle) {
-        patient.getColumnmaps().putAll(getAncCommonPersonObject(patient.entityId()).getColumnmaps());
-        startActivity(initProfileActivityIntent(patient, bundle, getAncMemberProfileActivityClass()));
-    }
-
-    public void gotToPncProfileActivity(CommonPersonObjectClient patient, Bundle bundle) {
-        patient.getColumnmaps().putAll(getPncCommonPersonObject(patient.entityId()).getColumnmaps());
-        startActivity(initProfileActivityIntent(patient, bundle, getPncMemberProfileActivityClass()));
-    }
-
-    private Intent initProfileActivityIntent(CommonPersonObjectClient patient, Bundle bundle, Class activityClass) {
-        HashMap<String, String> familyHeadDetails = getAncFamilyHeadNameAndPhone(getFamilyHead());
-        String familyHeadName = "";
-        String familyHeadPhone = "";
-        if (familyHeadDetails != null) {
-            familyHeadName = familyHeadDetails.get(org.smartregister.chw.anc.util.Constants.ANC_MEMBER_OBJECTS.FAMILY_HEAD_NAME);
-            familyHeadPhone = familyHeadDetails.get(org.smartregister.chw.anc.util.Constants.ANC_MEMBER_OBJECTS.FAMILY_HEAD_PHONE);
-        }
-        Intent intent = new Intent(this, activityClass);
-        if (bundle != null) {
-            intent.putExtras(bundle);
-        }
-        intent.putExtra(org.smartregister.chw.anc.util.Constants.ANC_MEMBER_OBJECTS.MEMBER_PROFILE_OBJECT, new MemberObject(patient));
-        intent.putExtra(org.smartregister.chw.anc.util.Constants.ANC_MEMBER_OBJECTS.FAMILY_HEAD_NAME, familyHeadName);
-        intent.putExtra(org.smartregister.chw.anc.util.Constants.ANC_MEMBER_OBJECTS.FAMILY_HEAD_PHONE, familyHeadPhone);
-        intent.putExtra(TITLE_VIEW_TEXT, String.format(getString(org.smartregister.chw.core.R.string.return_to_family_name), ""));
-        return intent;
     }
 
     @Override
